requirements.txt
asgiref==3.6.0
coverage==7.2.5
Django==4.2.1
factory-boy==3.2.1
Faker==18.7.0
flake8==6.0.0
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-decouple==3.8
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
